Fix `yarn test-e2e` and testcafe for single package.json structure
We're attempting to go to a single package.json structure, but the
testcafe setup was still implicitly referencing app/node_modules. This
adjusts testcafe so that it finds the electron browser (provided by
testcafe-browser-provider-electron).

Please double check this. I'm not familiar with testcafe, I just
messed with paths until I got the tests passing.